Add unboxer when key is passed
Problem: Encrypted blobs aren't being unboxed with the `?unbox=` query
parameter, which means that this problem is kicked down the road to the
application to manually unbox all blobs. This is a pain point when using
plain old `<img>` elements.

Solution: Copy the behavior from SSB-WS so that `?unbox=` query
parameters are used as unbox keys and decrypted before being passed
along. These modules solve many of the same problems, but the deep
Multiserver integration in SSB-WS makes it difficult to know which port
the server is running on. There's also Multiblob-HTTP, which seems to
have very similar behavior to SSB-Serve-Blobs, but it *also* doesn't
implement unboxing. Someday maybe these modules should be consolidated,
but for the time being I think change is a move in the right direction.